+++ src/v1/Helpers/resolveToContentType.spec.ts
@@ -31,41 +31,29 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { OnError, THROW_THE_ERROR } from "@ganbarodigital/ts-lib-error-reporting/lib/v1";
+import { expect } from "chai";
+import { describe } from "mocha";
 
-import { UnexpectedContentTypeError } from "../Errors/UnexpectedContentType";
-import { matchesContentType } from "./matchesContentType";
-import { resolveToContentType, ContentTypeOrMediaType } from "../Helpers";
+import { resolveToContentType } from ".";
+import { MediaType } from "../MediaType";
+import { contentTypeFrom } from "../ContentType";
 
-/**
- * Data guarantee. Calls your onError handler if the given input
- * doesn't match any of the MediaTypes on the given safelist.
- *
- * We compare everything except the parameters of the MediaTypes.
- */
-export function mustMatchContentType(
-    input: ContentTypeOrMediaType,
-    safelist: ContentTypeOrMediaType[],
-    onError: OnError = THROW_THE_ERROR,
-): void {
-    // does it match?
-    if (matchesContentType(input, safelist)) {
-        // yes it does!
-        return;
-    }
+describe("resolveToContentType()", () => {
+    it("converts a MediaType to a ContentType", () => {
+        const inputValue = new MediaType("text/html");
+        const expectedValue = "text/html";
 
-    // which content types have we checked it against?
-    const checkedAgainst = safelist.map(
-        (mt) => resolveToContentType(mt)
-    );
+        const actualValue = resolveToContentType(inputValue);
 
-    // tell the caller what happened
-    onError(new UnexpectedContentTypeError({
-        public: {
-            input: resolveToContentType(input),
-            required: {
-                anyOf: checkedAgainst,
-            },
-        }
-    }));
-}
+        expect(actualValue).to.equal(expectedValue);
+    });
+
+    it("returns a given ContentType", () => {
+        const inputValue = contentTypeFrom("text/html");
+        const expectedValue = inputValue;
+
+        const actualValue = resolveToContentType(inputValue);
+
+        expect(actualValue).to.equal(expectedValue);
+    })
+});