+++ src/v1/index.ts
@@ -33,4 +33,5 @@
 //
 
 export * from "./Errors";
+export * from "./ContentType";
 export * from "./MediaType";